+++ lib/netzke/basepack/accordion_panel.rb
@@ -2,24 +2,29 @@ module Netzke
   module Basepack
     # = AccordionPanel
     #
-    # A panel with the 'accordion' layout. Can lazily load its nested components. For example:
+    # A panel with the 'accordion' layout. By default, lazily loads its nested components. For example:
     #
-    #     netzke :my_accordion, :items => [{
+    #   class MyAccordion < Netzke::Basepack::AccordionPanel
+    #     def items
+    #       [{
+    #         # just an Ext panel
     #         :html => "I'm a simple Ext.Panel",
     #         :title => "Panel One"
     #       },{
-    #         :class_name => "SimplePanel",
+    #         # a Netzke component
+    #         :klass => SimplePanel,
     #         :update_text => "Update for Panel Two",
-    #         :title => "Panel Two",
-    #         :lazy_loading => true
+    #         :title => "Panel Two"
     #       }]
+    #     end
+    #   end
     class AccordionPanel < Netzke::Base
 
       include WrapLazyLoaded
 
       js_configure do |c|
         c.layout = :accordion
-        c.component_load_mask = {:msg => "null".l} # due to a probable bug in Ext's Accordion Layout (mask message is mis-layed-out), disabling mask message
+        # c.component_load_mask = {:msg => "null".l} # due to a probable bug in Ext's Accordion Layout (mask message is mis-layed-out), disabling mask message
 
         c.init_component = <<-JS
           function(params){