+++ lib/netzke/basepack/wrap_lazy_loaded.rb
@@ -5,24 +5,23 @@ module Netzke
     # Used, for instance, in TabPanel and AccordionPanel to dynamically load components on expanding a panel or clicking
     # a tab.
     module WrapLazyLoaded
-      def js_items
-        orig = super.dup
-        orig.each do |item|
-          wrapped_component = components[item[:netzke_component]]
+      def extend_item(item)
+        item = super
 
-          # When a nested component with lazy loading is detected, it gets replaced with a 'fit' panel,
-          # into which later the component itself is dynamically loaded on request.
-          if wrapped_component && wrapped_component[:lazy_loading]
-            item.replace({
-              :layout => 'fit',
-              :wrapped_component => wrapped_component[:item_id],
-              :title => wrapped_component[:title] || wrapped_component[:item_id],
-              :icon_cls => wrapped_component[:icon_cls],
-              :disabled => wrapped_component[:disabled]
-            })
-          end
+        # when a nested component with lazy loading is detected, it gets replaced with a 'fit' panel,
+        # into which later the component itself is dynamically loaded on request.
+        merged_config = components[item[:netzke_component]].try(:merge, item)
+        if merged_config && !merged_config[:eager_loading]
+          {
+            layout: :fit,
+            wrapped_component: merged_config[:item_id],
+            title: merged_config[:title] || merged_config[:item_id],
+            icon_cls: merged_config[:icon_cls],
+            disabled: merged_config[:disabled]
+          }
+        else
+          item
         end
-        orig
       end
     end
   end