+++ test/basepack_test_app/app/components/some_accordion.rb
@@ -1,18 +1,13 @@
-class SomeTabPanel < Netzke::Basepack::TabPanel
+class SomeAccordion < Netzke::Basepack::AccordionPanel
   # This component will be dynamically loaded on expanding the second accordion pane
   component :simple_panel do |c|
     c.update_text = "Update for Panel Two"
-    c.title = "Panel Two"
-    c.prevent_header = true
+    c.title = "Panel Two!"
     c.border = false
-
-    # optionally, you can force a certain component to be eagerly loaded:
-    # c.eager_loading = true
+    c.prevent_header = true
   end
 
   def configure(c)
-    c.active_tab = 0
-
     c.items = [
       { :html => "I'm a simple Ext.Panel", :title => "Panel One" },
       :simple_panel